Potential fix for code scanning alert no. 94

+++ tools/mkeficapsule.c
@@ -147,7 +147,7 @@ static int read_bin_file(char *bin, uint8_t **data, off_t *bin_size)
 
 	size = fread(buf, 1, bin_stat.st_size, g);
 	if (size < bin_stat.st_size) {
-		fprintf(stderr, "read failed (%zx)\n", size);
+		fprintf(stderr, "read failed (%zx)\n", (size_t)size);
 		ret = -1;
 		goto err;
 	}
